[Ref] Silence 2 signed/unsigned comparison warnings.
git-svn-id: https://source.openmpt.org/svn/openmpt/trunk/OpenMPT@2460 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ soundlib/Load_ams.cpp
@@ -184,7 +184,7 @@ void ReadAMSPattern(CPattern &pattern, bool newVersion, FileReader &patternChunk
 							}
 							break;
 						}
-					} else if(effect - 0x10 < CountOf(effTrans))
+					} else if(effect - 0x10 < (int)CountOf(effTrans))
 					{
 						// Extended commands
 						m.command = effTrans[effect - 0x10];
